requirements.txt
asgiref==3.5.1
certifi==2021.10.8
Django==4.0.4
django-filter==21.1
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
sentry-sdk==1.5.11
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
